pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "BACE"
version = "0.1.0"
description = "Bayesian Anchored Co-Evolution of Code and Test Populations"
readme = "README.md"
requires-python = "==3.12.*"

# List all external dependencies here
dependencies = [
    "notebook==7.4.7",
    "pandas==2.3.3",
    "ollama==0.6.0",
    "langchain==0.3.27",
    "openai==1.109.1",
    "langgraph==0.6.8",
    "fire==0.7.1",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "appnope==0.1.4",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "attrs==25.3.0",
    "babel==2.17.0",
    "beautifulsoup4==4.14.2",
    "bleach==6.2.0",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "comm==0.2.3",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "executing==2.2.1",
    "fastjsonschema==2.21.2",
    "fqdn==1.5.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.30.1",
    "ipython==9.6.0",
    "ipython-pygments-lexers==1.1.1",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.11.0",
    "json5==0.12.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.3.0",
    "jupyter-server==2.17.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.9",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "langchain-core==0.3.76",
    "langchain-text-splitters==0.3.11",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.4",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.31",
    "lark==1.3.0",
    "markupsafe==3.0.3",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.4",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "notebook-shim==0.2.4",
    "numpy==2.3.3",
    "orjson==3.11.3",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "pandocfilters==1.5.1",
    "parso==0.8.5",
    "pexpect==4.9.0",
    "platformdirs==4.4.0",
    "prometheus-client==0.23.1",
    "prompt-toolkit==3.0.52",
    "psutil==7.1.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycparser==2.23",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-json-logger==3.3.0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "referencing==0.36.2",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rfc3987-syntax==1.1.0",
    "rpds-py==0.27.1",
    "send2trash==1.8.3",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "sqlalchemy==2.0.43",
    "stack-data==0.6.3",
    "tenacity==9.1.2",
    "termcolor==3.1.0",
    "terminado==0.18.1",
    "tinycss2==1.4.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "tzdata==2025.2",
    "uri-template==1.3.0",
    "urllib3==2.5.0",
    "wcwidth==0.2.14",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "xxhash==3.5.0",
    "zstandard==0.25.0",
    "huggingface-hub>=0.35.3",
    "datasets==3.6.0",
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2",
    "loguru>=0.7.3",
    "typer>=0.20.0",
    "networkx>=3.5",
    "adjusttext>=1.3.0",
    "pytest-timeout>=2.3.1",
    "pyvis>=0.3.2",
    "transformers>=4.57.6",
    "torch>=2.10.0",
    "google-genai>=1.68.0",
    "google-auth>=2.49.1",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "pandas-stubs>=2.3.2.250926",
    "types-python-dateutil==2.9.0.20250822",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "types-pyyaml>=6.0.12.20250915",
    "types-psutil>=7.2.2.20260130",
]

[tool.setuptools.packages.find]
# This is the key part: it tells the installer to look for packages inside the 'src' directory
where = ["src"]


[tool.mypy]
strict = true
mypy_path = "src"
explicit_package_bases = true


[[tool.mypy.overrides]]
module = "fire"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "networkx.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (select with '-m integration')",
]
